Modularizacion completada, haciendo lo de nedb
