+++ index.js
@@ -1,5 +1,6 @@
-const express = require("express");
-const cool = require('cool-ascii-faces');
+import express from "express";
+import {loadBackendGAM} from "./src/back/index-GAM.js";
+//const cool = require('cool-ascii-faces');
 const app = express();
 const PORT = process.env.PORT || 16078;
 const BASE_API= "/api/v1";
@@ -8,205 +9,14 @@ app.use("/",express.static("./public"));
 app.use("/about",express.static("./public/about.html"));
 app.use(express.json());
 
-app.get("/cool",(request,response)=>{
-    response.send(cool());
-});
+// app.get("/cool",(request,response)=>{
+//     response.send(cool());
+// });
+
+loadBackendGAM(app);
 
 app.listen(PORT,()=>{
     console.log(`Server running on port ${PORT}...`);
-});
-
-
-// index-GAM.js
-
-const initialEmigrationData = [
-    {autonomic_community: "andalucia", year: 2021, quarter: "q1", between_20_24_yo: 3666, between_25_29_yo: 5409, between_30_34_yo: 5996 },
-    { autonomic_community: "andalucia", year: 2020, quarter: "q2", between_20_24_yo: 2156, between_25_29_yo: 3201, between_30_34_yo: 3690 },
-    {autonomic_community: "asturias", year: 2021, quarter: "q3", between_20_24_yo: 304, between_25_29_yo: 510, between_30_34_yo: 483 },
-    { autonomic_community: "islas-baleares", year: 2021, quarter: "q1", between_20_24_yo: 6320, between_25_29_yo: 1023, between_30_34_yo: 1239 },
-    { autonomic_community: "canarias", year: 2021, quarter: "q2", between_20_24_yo: 947, between_25_29_yo: 1625, between_30_34_yo: 1643 },
-    { autonomic_community: "madrid", year: 2021, quarter: "q3", between_20_24_yo: 6028, between_25_29_yo: 10836, between_30_34_yo: 10004 },
-    { autonomic_community: "castilla-y-leon", year: 2021, quarter: "q1", between_20_24_yo: 766, between_25_29_yo: 1171, between_30_34_yo: 1210 },
-    { autonomic_community: "castilla-la-mancha", year: 2021, quarter: "q2", between_20_24_yo: 984, between_25_29_yo: 1304, between_30_34_yo: 1512 },
-    { autonomic_community: "cataluña", year: 2021, quarter: "q3", between_20_24_yo: 7305, between_25_29_yo: 12960, between_30_34_yo: 13077 },
-    { autonomic_community: "cataluña", year: 2020, quarter: "q1", between_20_24_yo: 4469, between_25_29_yo: 8086, between_30_34_yo: 7808 },
-    { autonomic_community: "cataluña", year: 2019, quarter: "q1", between_20_24_yo: 6397, between_25_29_yo: 12400, between_30_34_yo: 12023 },
-    { autonomic_community: "madrid", year: 2020, quarter: "q1", between_20_24_yo: 3981, between_25_29_yo: 6753, between_30_34_yo: 6239 }
-];
-
-const emigrationData=new Array();
-
-let array_between_30_34_yo_cat= initialEmigrationData.slice(-3).map(obj=>
-    obj.between_30_34_yo
-);
-
-function average(){
-    let acc=0;
-    if(array_between_30_34_yo_cat.length===0){
-        return 0;
-    }
-    array_between_30_34_yo_cat.forEach(n =>{
-        acc+= n;
-    });
-
-    return acc/array_between_30_34_yo_cat.length;
-}
-
-app.get("/samples/GAM",(request,response)=>{
-    let res= average();
-    response.send(res.toString());
-});
-
-//11.
-
-app.get(BASE_API+"/emigration-stats",(request,response)=>{
-    let res= emigrationData;
-    response.send(JSON.stringify(res,null,2));
-});
-
-//13.
-
-app.get(BASE_API + "/emigration-stats/loadInitialData", (request,response) =>{
-    if(!emigrationData.length){
-        emigrationData.push(...initialEmigrationData);
-    }
-    response.send(initialEmigrationData);
-})
-
-
-//16.a.1
-app.get(BASE_API+"/emigration-stats/:name",(request,response)=>{
-    let paramName = request.params.name;
-    console.log(`New GET to /emigration-stats/${paramName}`);
-    //let id= Number(request.query.id);
-    /*if (id){
-        let res= emigrationData.filter(obj => obj.id === id); // busca por id
-        response.send(JSON.stringify(res,null,2));
-    }*/
-    let res= emigrationData.filter(obj => obj.autonomic_community === paramName);
-    if(res.length === 0){
-        response.sendStatus(404);
-    }
-    response.send(JSON.stringify(res,null,2));
-});
-
-//duda : nose si el /towns ponerlo por /comunidad_autonoma o dejarlo con todas y luego el especifico la verdad xd
-
-app.post(BASE_API+"/emigration-stats",(request,response)=>{ 
-    console.log("New POST to /emigration-stats");
-    console.log(`<${request.body}>`); // <> para saber si esta vacio
-
-    const allowedFields = ["autonomic_community", "year", "quarter", "between_20_24_yo", "between_25_29_yo", "between_30_34_yo"];
-    let newAutonomicCommunity=request.body;
-    let invalidFields= Object.keys(newAutonomicCommunity).filter(f => !allowedFields.includes(f))
-    if(invalidFields.length>0){ // si funca
-        response.sendStatus(400);
-    }
-    //let emigrationDataWithoutId=emigrationData.map(({id, ...resto}) => resto );
-    else if(emigrationData.some(i => JSON.stringify(i) === JSON.stringify(newAutonomicCommunity))){ // veo si ya existe por el nuevoId (mala idea, el id se autoincrementa), el some devuelve booleano, no funca
-        response.sendStatus(409); // copiar la primera haceiendo un slice sin el id o borrar el id
-    }
-    else{
-        response.sendStatus(201);
-        emigrationData.push(newAutonomicCommunity);
-    }
-
-    //let lastId=emigrationData[emigrationData.length -1].id;
-    //let newId=lastId+1;
-    //newAutonomicCommunity= {id: newId, ...newAutonomicCommunity};
-});
-
-app.put(BASE_API+"/emigration-stats",(request,response)=>{ // método incorrecto
-    response.sendStatus(405);
-});
-
-app.delete(BASE_API+"/emigration-stats",(request,response)=>{ // dudas, dejarlo asi y cambiarlo el dia del feedback
-    emigrationData.length=0; // la vacio
-    response.send(200);
-});
-
-//16.a.2
-
-app.get(BASE_API+"/emigration-stats/:name/:year/:quarter",(request,response)=>{
-
-    //let id= Number(request.query.id);
-    // if (id){
-    //     let res= emigrationData.filter(obj => obj.id === id); // busca por id
-    //     response.send(JSON.stringify(res,null,2));
-    // }
-    let paramName = request.params.name;
-    let paramYear = request.params.year;
-    let paramQuarter = request.params.quarter;
-    console.log(`New GET to /emigration-stats/${paramName}/${paramYear}/${paramQuarter}`);
-    let res = emigrationData.filter(v => v.autonomic_community === paramName && parseInt(v.year)===parseInt(paramYear) && v.quarter === paramQuarter);
-    if(res.length===0){
-        response.sendStatus(404);
-    }
-    response.send(JSON.stringify(res,null,2));
-});
-
-app.post(BASE_API+"/emigration-stats/:name/:year/:quarter",(request,response)=>{
-    response.sendStatus(405);
-});
-
-app.put(BASE_API+"/emigration-stats/:name/:year/:quarter",(request,response)=>{ // dudas, actualizo todas las de cataluña? o solo una en especifico (id ?)
-    //let id= Number(request.query.id); // de la URL, hay que parsearlo aqui o despues porque sale como String de la URL
-    //let {id :bodyId, ...updatedData } = request.body; // del Body
-    let paramName = request.params.name; // comprobar que es igual al del body, decido que es ese el identifiacador (parametro obligatorio en el body)
-    let paramYear = request.params.year;
-    let paramQuarter = request.params.quarter;
-    let putBody= request.body;
-    //let {autonomic_community:bodyName, ...updatedData} = request.body;
-    let allowedFields = ["autonomic_community", "year", "quarter", "between_20_24_yo", "between_25_29_yo", "between_30_34_yo"];
-    let invalidFields = Object.keys(putBody).filter(f => !allowedFields.includes(f));
-    //let ind=emigrationData.findIndex(i => i.autonomic_community === paramName && i.year === paramYear && i.quarter===paramQuarter );
-    if(invalidFields.length>0){
-        response.sendStatus(400);
-    }
-    else if(!((putBody.autonomic_community === paramName) && (parseInt(putBody.year) === parseInt(paramYear)) && (putBody.quarter === paramQuarter))){
-        response.sendStatus(400);
-    }
-    else if (emigrationData.filter(v => v.autonomic_community === paramName && parseInt(v.year)===parseInt(paramYear) && v.quarter === paramQuarter).length == 0){
-        response.sendStatus(404);
-    }
-    else{
-        emigrationData.forEach(v => {
-            if((v.autonomic_community === paramName && parseInt(v.year)===parseInt(paramYear) && v.quarter === paramQuarter)){
-                v.between_20_24_yo = putBody.between_20_24_yo;
-                v.between_25_29_yo = putBody.between_25_29_yo;
-                v.between_30_34_yo = putBody.between_30_34_yo;
-            }
-        })
-        response.sendStatus(200);
-
-    }
-
-    // if(Number(bodyId) != id){ 
-    //     response.sendStatus(400); // que debe aparecer el id en el body de la peticion, pero tambien asegurarme de que parazca en la peticion?
-    // }
-    // let ind=emigrationData.findIndex(i => i.id === id); // devuelve la posicion del array en la que se encuentra el id
-    // if( ind === -1){
-    //     response.sendStatus(404); // igual hay que chequear que esta dentro de los id que contienen a cataluña o hacer un slice, que los reenumere con id del 0 al ultimo tambien
-    // }
-    // emigrationData[ind] = {...emigrationData[ind], ...updatedData};
-    // response.sendStatus(200);
-});
-
-app.delete(BASE_API+"/emigration-stats/:name/:year/:quarter",(request,response)=>{ // dudas, borro todas las de cataluña? o solo una en especifico (id ?)
-    let paramName = request.params.name;
-    let paramYear = request.params.year;
-    let paramQuarter = request.params.quarter;
-    console.log(`New DELETE to /emigration-stats/${paramName}/${paramYear}/${paramQuarter}`);
-    let res = emigrationData.filter(v => v.autonomic_community === paramName && parseInt(v.year) === parseInt(paramYear) && v.quarter === paramQuarter);
-
-    if(request.length===0){
-        response.sendStatus(404);
-    } else{
-        console.log(res)
-        res.forEach((item) => emigrationData.splice(emigrationData.indexOf(item), 1));
-    }
-
-    response.sendStatus(200);
 });
 
 // index-PVS.js
